Make Strings Copy
I'd like to try to make documents as Copy as possible, so strings are
now stored in a context and extracted as needed.

It's like interning but more generic.

In the future, I may try to make Structures Copy (like pretty.rs' arena
approach), which would make them much easier to manage, at the cost of
an extra generic (but that generic was already added for string
interning, perhaps).